alembic: upgrade funders recipe
